requirement.txt
tensorflow
flask
numpy
Pillow
flasgger
gevent
faskapi
pydantic
scipy